OCT-1667 & OCT-1668: Implement QF sliding scale and PoU for simmulations

+++ backend/app/infrastructure/database/uniqueness_quotient.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Optional
 
 from app import db
@@ -20,7 +21,7 @@ def get_uq_by_address(user_address: str, epoch: int) -> Optional[UniquenessQuoti
     return get_uq_by_user(user, epoch)
 
 
-def save_uq(user: User, epoch: int, score: float):
+def save_uq(user: User, epoch: int, score: Decimal):
     uq: UniquenessQuotient = UniquenessQuotient(
         epoch=epoch,
         user_id=user.id,
